rm NATIVE_ASSETS_PER_CHAIN
